requirements.txt
pillow>=10.0.0
pillow-heif>=0.13.0
exifread>=3.0.0
ffmpeg-python>=0.2.0
python-dateutil>=2.8.0
